- refactor api base url and image base url
