+++ lib/core/networking/api_constants.dart
@@ -1,6 +1,6 @@
 class ApiConstants {
-  static const String apiBaseUrl = "http://home4u.gosoftcloud.com";
-  static const String imageBaseUrl = "http://149.102.135.15:5001/";
+  static const String apiBaseUrl = "https://home4u.gosoftcloud.com";
+  static const String imageBaseUrl = "https://149.102.135.15:5001/";
   static const String loginEndPoint = "/api/v1/auth/login";
   static const String userTypesEp = "/api/v1/user-types";
   static const String signUpEp = "/api/v1/auth/register";
@@ -48,7 +48,7 @@ class ApiConstants {
 
   ///images
   static String getImageBaseUrl(String path) =>
-      "$imageBaseUrl$path";
+      "$apiBaseUrl/$path";
 }
 
 class ApiErrors {
