+++ lib/features/profile/presentation/widgets/projects_widgets/projects_details_shimmer_widget.dart
@@ -68,8 +68,9 @@ class _ProjectsDetailsShimmerWidgetState
                   children: [
                     CircleAvatar(
                       radius: 20,
-                      backgroundImage: NetworkImage(
-                          'https://media.istockphoto.com/id/1142192548/vector/man-avatar-profile-male-face-silhouette-or-icon-isolated-on-white-background-vector.jpg?s=2048x2048&w=is&k=20&c=lyki7QHyULuJNNheEf-BI_DQNCDi2NRYMfVGTQj_4UM='),
+                      // backgroundImage: NetworkImage(
+                      //     'https://media.istockphoto.com/id/1142192548/vector/man-avatar-profile-male-face-silhouette-or-icon-isolated-on-white-background-vector.jpg?s=2048x2048&w=is&k=20&c=lyki7QHyULuJNNheEf-BI_DQNCDi2NRYMfVGTQj_4UM='),
+                      //
                     ),
                     horizontalSpace(16),
                     Text(
